UI overhaul.
The tasks will be shown under the tasks url instead of the root.

The tasks are now shown using URL params and a separate NextJS page instead of dialogs.

Tasks are now created by entering the title at the top of tasks list and pressing enter.

Tasks are now passed to child components using context instead of props.

Removes filters(Since it only had 2 options it) and now the app shows pending tasks at the top and  completed tasks at the bottom.