Beginning of v2
I've decided to make this app usable. To do that I've decided to use
Postgres with Drizzle instead MongoDB with Prisma(This one is for my
ease of development, so it has nothing to do with app). I'm also
overhauling the UI(Kanban).

This is the first commit of a long series of commits.