Version 2 (#9)
* Beginning of v2

I've decided to make this app usable. To do that I've decided to use
Postgres with Drizzle instead MongoDB with Prisma(This one is for my
ease of development, so it has nothing to do with app). I'm also
overhauling the UI(Kanban).

This is the first commit of a long series of commits.

* Adds a landing page

* Adds a working Kanban board

* Adds sidebar which shows list of boards created by the user.

* Boards and tasks are saved and fetched from the DB

* Boards can now be pinned or updated

* Column can now be created, re-ordered and deleted.

* Adds packageManager in package.json for vercel to use corepack.

* Improvements to login screen.

* Tasks can be now moved within and to other columns.

* Add task editing and deletion functionality

Implement task edit modal with title/description fields, optimize column operations with Promise.all.
imports/unused code. Add task deletion API endpoint and database support.